refactor(api): rename attestation and certificate PDF generation files
