+++ api/src/certification/results/application/certificate-controller.js
@@ -8,8 +8,8 @@ import { usecases } from '../domain/usecases/index.js';
 import * as certificateSerializer from '../infrastructure/serializers/certificate-serializer.js';
 import * as certificateSummarySerializer from '../infrastructure/serializers/certificate-summary-serializer.js';
 import * as privateCertificateSerializer from '../infrastructure/serializers/private-certificate-serializer.js';
-import * as v3CertificationAttestationPdf from '../infrastructure/utils/pdf/generate-pdf-certificate.js';
-import * as v2CertificationAttestationPdf from '../infrastructure/utils/pdf/generate-v2-pdf-certificate.js';
+import * as v2CertificationAttestationPdf from '../infrastructure/utils/pdf/generate-v2-pdf-attestation.js';
+import * as v3CertificationAttestationPdf from '../infrastructure/utils/pdf/generate-v3-pdf-certificate.js';
 
 async function getCertificateByVerificationCode(request, h, dependencies = { certificateSerializer }) {
   const locale = getChallengeLocale(request);
