+++ api/src/certification/results/infrastructure/utils/pdf/templates/v2-attestation.js
