+++ api/tests/certification/results/integration/infrastructure/utils/pdf/generate-v3-pdf-certificate_test.js
@@ -4,7 +4,7 @@ import url from 'node:url';
 import dayjs from 'dayjs';
 import { getDocument } from 'pdfjs-dist/legacy/build/pdf.mjs';
 
-import { generate } from '../../../../../../../src/certification/results/infrastructure/utils/pdf/generate-pdf-certificate.js';
+import { generate } from '../../../../../../../src/certification/results/infrastructure/utils/pdf/generate-v3-pdf-certificate.js';
 import { Frameworks } from '../../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { getI18n } from '../../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { expect } from '../../../../../../test-helper.js';
@@ -12,7 +12,7 @@ import { domainBuilder } from '../../../../../../tooling/domain-builder/domain-b
 
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
-describe('Integration | Infrastructure | Utils | Pdf | V3 Certification Attestation Pdf', function () {
+describe('Integration | Infrastructure | Utils | Pdf | V3 Certificate Pdf', function () {
   let i18n, translate;
 
   beforeEach(function () {
